refactor: Coordinate Common 으로 패키지 변경
