+++ backend/pcloud-domain/src/testFixtures/java/show/exhibition/domain/ExhibitionFixture.java
@@ -1,15 +1,16 @@
 package show.exhibition.domain;
 
+import com.domain.common.coordinates.Latitude;
+import com.domain.common.coordinates.Longitude;
+import com.domain.common.coordinates.Position;
 import com.domain.show.common.Price;
 import com.domain.show.common.PublicTag;
 import com.domain.show.common.ShowDetails;
 import com.domain.show.common.ShowRules;
 import com.domain.show.common.ShowSchedule;
-import com.domain.show.exhibition.domain.Exhibition;
-import com.domain.show.common.Position;
-import com.domain.show.common.Latitude;
-import com.domain.show.common.Longitude;
 import com.domain.show.common.Statistic;
+import com.domain.show.exhibition.domain.Exhibition;
+
 import java.time.LocalDateTime;
 
 @SuppressWarnings("NonAsciiCharacters")
